+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/tournaments/TournamentsService.java
@@ -1,13 +1,16 @@
 package de.mephisto.vpin.server.tournaments;
 
-import com.google.common.annotations.VisibleForTesting;
 import de.mephisto.vpin.commons.fx.Features;
 import de.mephisto.vpin.connectors.mania.VPinManiaClient;
+import de.mephisto.vpin.connectors.mania.model.Cabinet;
 import de.mephisto.vpin.restclient.PreferenceNames;
 import de.mephisto.vpin.restclient.tournaments.TournamentConfig;
 import de.mephisto.vpin.restclient.tournaments.TournamentSettings;
 import de.mephisto.vpin.restclient.util.SystemUtil;
+import de.mephisto.vpin.server.highscores.HighscoreService;
+import de.mephisto.vpin.server.preferences.PreferenceChangedListener;
 import de.mephisto.vpin.server.preferences.PreferencesService;
+import de.mephisto.vpin.server.system.SystemService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
@@ -19,7 +22,7 @@ import org.springframework.web.server.ResponseStatusException;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 
 @Service
-public class TournamentsService implements InitializingBean {
+public class TournamentsService implements InitializingBean, PreferenceChangedListener {
   private final static Logger LOG = LoggerFactory.getLogger(TournamentsService.class);
 
   @Value("${vpinmania.server.host}")
@@ -27,6 +30,15 @@ public class TournamentsService implements InitializingBean {
 
   @Autowired
   private PreferencesService preferencesService;
+
+  @Autowired
+  private HighscoreService highscoreService;
+
+  @Autowired
+  private SystemService systemService;
+
+  private VPinManiaClient maniaClient;
+  private TournamentsHighscoreChangeListener tournamentsHighscoreChangeListener;
 
   public TournamentConfig getConfig() {
     TournamentConfig config = new TournamentConfig();
@@ -51,6 +63,25 @@ public class TournamentsService implements InitializingBean {
 
   @Override
   public void afterPropertiesSet() throws Exception {
+    if (Features.TOURNAMENTS_ENABLED) {
+      TournamentConfig config = getConfig();
+      maniaClient = new VPinManiaClient(config.getUrl(), config.getSystemId());
 
+      tournamentsHighscoreChangeListener = new TournamentsHighscoreChangeListener(maniaClient);
+      highscoreService.addHighscoreChangeListener(tournamentsHighscoreChangeListener);
+
+      preferencesService.addChangeListener(this);
+      preferenceChanged(PreferenceNames.TOURNAMENTS_SETTINGS, null, null);
+    }
+  }
+
+  @Override
+  public void preferenceChanged(String propertyName, Object oldValue, Object newValue) throws Exception {
+    if (PreferenceNames.TOURNAMENTS_SETTINGS.equals(propertyName)) {
+      Cabinet cabinet = maniaClient.getCabinetClient().getCabinet();
+      this.tournamentsHighscoreChangeListener.setCabinet(cabinet);
+
+      LOG.info("Registered Tournaments HighscoreChangeListener");
+    }
   }
 }