+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/preferences/AvatarUploadProgressModel.java
@@ -1,14 +1,18 @@
 package de.mephisto.vpin.ui.preferences;
 
 import de.mephisto.vpin.commons.fx.Features;
+import de.mephisto.vpin.connectors.mania.model.Tournament;
 import de.mephisto.vpin.ui.util.ProgressModel;
 import de.mephisto.vpin.ui.util.ProgressResultModel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.util.Arrays;
 import java.util.Iterator;
+import java.util.List;
 
 import static de.mephisto.vpin.ui.Studio.client;
 import static de.mephisto.vpin.ui.Studio.maniaClient;
@@ -58,7 +62,14 @@ public class AvatarUploadProgressModel extends ProgressModel<File> {
     try {
       client.getPreferenceService().uploadVPinAvatar(f);
       if (Features.TOURNAMENTS_ENABLED && maniaClient.getCabinetClient() != null) {
-        maniaClient.getCabinetClient().updateAvatar(f, null);
+        maniaClient.getCabinetClient().updateAvatar(f, null, true);
+
+        BufferedImage avtr = ImageIO.read(f);
+        List<Tournament> tournaments = maniaClient.getTournamentClient().getTournaments();
+        for (Tournament tournament : tournaments) {
+          maniaClient.getTournamentClient().updateBadge(avtr, tournament);
+          LOG.info("Updated tournament badge for " + tournament);
+        }
       }
     } catch (Exception e) {
       LOG.error("Error waiting for server: " + e.getMessage(), e);