+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/TournamentCreationProgressModel.java
@@ -2,32 +2,35 @@ package de.mephisto.vpin.ui.tournaments;
 
 import de.mephisto.vpin.connectors.mania.model.Account;
 import de.mephisto.vpin.connectors.mania.model.Tournament;
+import de.mephisto.vpin.connectors.mania.model.TournamentTable;
+import de.mephisto.vpin.ui.tournaments.view.TournamentTreeModel;
 import de.mephisto.vpin.ui.util.ProgressModel;
 import de.mephisto.vpin.ui.util.ProgressResultModel;
-import org.apache.commons.io.IOUtils;
+import javafx.scene.control.TreeItem;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.FileOutputStream;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 
 import static de.mephisto.vpin.ui.Studio.maniaClient;
 
-public class TournamentCreationProgressModel extends ProgressModel<Tournament> {
+public class TournamentCreationProgressModel extends ProgressModel<TreeItem<TournamentTreeModel>> {
   private final static Logger LOG = LoggerFactory.getLogger(TournamentCreationProgressModel.class);
-  private List<Tournament> tournaments;
+  private final List<TreeItem<TournamentTreeModel>> tournaments;
   private final Account account;
   private final BufferedImage badge;
+  private Tournament newTournament;
 
-  private Iterator<Tournament> iterator;
+  private final Iterator<TreeItem<TournamentTreeModel>> iterator;
 
-  public TournamentCreationProgressModel(Tournament tournament, Account account, BufferedImage badge) {
-    super("Creating Tournament \"" + tournament.getDisplayName() + "\"");
-    this.tournaments = Arrays.asList(tournament);
+  public TournamentCreationProgressModel(TreeItem<TournamentTreeModel> tournamentModel, Account account, BufferedImage badge) {
+    super("Creating Tournament \"" + tournamentModel.getValue().getTournament().getDisplayName() + "\"");
+    this.tournaments = new ArrayList<>(Arrays.asList(tournamentModel));
+    this.tournaments.addAll(tournamentModel.getChildren());
     this.account = account;
     this.badge = badge;
     this.iterator = this.tournaments.iterator();
@@ -54,20 +57,30 @@ public class TournamentCreationProgressModel extends ProgressModel<Tournament> {
   }
 
   @Override
-  public Tournament getNext() {
+  public TreeItem<TournamentTreeModel> getNext() {
     return iterator.next();
   }
 
   @Override
-  public String nextToString(Tournament game) {
-    return "Saving Tournament";
+  public String nextToString(TreeItem<TournamentTreeModel> item) {
+    if(item.getValue().isTournamentNode()) {
+      return "Saving Tournament \"" + item.getValue().getTournament().getDisplayName() + "\"";
+    }
+    return "Adding \"" + item.getValue().getVpsTable().getDisplayName() + "\"";
   }
 
   @Override
-  public void processNext(ProgressResultModel progressResultModel, Tournament next) {
+  public void processNext(ProgressResultModel progressResultModel, TreeItem<TournamentTreeModel> next) {
     try {
-      Tournament tournament = maniaClient.getTournamentClient().create(next, account, badge);
-      progressResultModel.getResults().add(tournament);
+      if(next.getValue().isTournamentNode()) {
+        newTournament = maniaClient.getTournamentClient().create(next.getValue().getTournament(), account, badge);
+        progressResultModel.getResults().add(newTournament);
+      }
+      else {
+        TournamentTable tournamentTable = next.getValue().getTournamentTable();
+        tournamentTable.setTournamentId(newTournament.getId());
+        maniaClient.getTournamentClient().addTable(tournamentTable);
+      }
     } catch (Exception e) {
       LOG.error("Error creating tournament: " + e.getMessage(), e);
       progressResultModel.getResults().add(e.getMessage());