+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/view/TournamentCellContainer.java
@@ -1,6 +1,7 @@
 package de.mephisto.vpin.ui.tournaments.view;
 
 import de.mephisto.vpin.connectors.mania.model.Tournament;
+import de.mephisto.vpin.connectors.mania.model.TournamentTable;
 import de.mephisto.vpin.ui.Studio;
 import de.mephisto.vpin.ui.util.AvatarFactory;
 import javafx.scene.control.Label;
@@ -8,6 +9,7 @@ import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 
 import java.io.InputStream;
+import java.util.List;
 
 import static de.mephisto.vpin.ui.Studio.client;
 import static de.mephisto.vpin.ui.Studio.maniaClient;
@@ -40,7 +42,9 @@ public class TournamentCellContainer extends HBox {
     titleLabel = new Label("Tables:");
     titleLabel.setPrefWidth(TITLE_WIDTH);
     titleLabel.getStyleClass().add("default-headline");
-    valueLabel = new Label(String.valueOf(tournament.getTableIdList().size()));
+
+    List<TournamentTable> tournamentTables = maniaClient.getTournamentClient().getTournamentTables(tournament.getId());
+    valueLabel = new Label(String.valueOf(tournamentTables.size()));
     valueLabel.getStyleClass().add("default-text");
     row.getChildren().addAll(titleLabel, valueLabel);
     entries.getChildren().add(row);