+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/view/TournamentTreeModel.java
@@ -1,11 +1,12 @@
 package de.mephisto.vpin.ui.tournaments.view;
 
 import de.mephisto.vpin.connectors.mania.model.Tournament;
+import de.mephisto.vpin.connectors.mania.model.TournamentTable;
 import de.mephisto.vpin.connectors.vps.VPS;
 import de.mephisto.vpin.connectors.vps.model.VpsTable;
 import de.mephisto.vpin.connectors.vps.model.VpsTableVersion;
-import de.mephisto.vpin.restclient.highscores.ScoreSummaryRepresentation;
 import de.mephisto.vpin.restclient.games.GameRepresentation;
+import de.mephisto.vpin.restclient.highscores.ScoreSummaryRepresentation;
 import de.mephisto.vpin.ui.Studio;
 import javafx.scene.control.TreeItem;
 import org.apache.commons.lang3.StringUtils;
@@ -16,49 +17,42 @@ import static de.mephisto.vpin.ui.Studio.client;
 
 public class TournamentTreeModel {
   private final Tournament tournament;
+  private final TournamentTable tournamentTable;
   private final VpsTableVersion vpsTableVersion;
   private final VpsTable vpsTable;
   private final GameRepresentation game;
 
   private boolean highscoreAvailable;
 
-  public static void expandTreeView(TreeItem<?> item){
-    if(item != null && !item.isLeaf()){
+  public static void expandTreeView(TreeItem<?> item) {
+    if (item != null && !item.isLeaf()) {
       item.setExpanded(true);
-      for(TreeItem<?> child:item.getChildren()){
+      for (TreeItem<?> child : item.getChildren()) {
         expandTreeView(child);
       }
     }
   }
 
-  public static TreeItem<TournamentTreeModel> create(Tournament tournament) {
-    TreeItem<TournamentTreeModel> tournamentNode = new TreeItem<>(new TournamentTreeModel(tournament, null, null, null));
-    List<String> tableIdList = tournament.getTableIdList();
-    for (String s : tableIdList) {
+  public static TreeItem<TournamentTreeModel> create(Tournament tournament, List<TournamentTable> tournamentTables) {
+    TreeItem<TournamentTreeModel> tournamentNode = new TreeItem<>(new TournamentTreeModel(tournament, null, null, null, null));
+    for (TournamentTable tournamentTable : tournamentTables) {
+      VpsTable table = VPS.getInstance().getTableById(tournamentTable.getVpsTableId());
       VpsTableVersion version = null;
-      VpsTable table = null;
       GameRepresentation game = null;
-      String[] split = s.split("#");
-      if (split.length > 0) {
-        String tableId = split[0];
-        table = VPS.getInstance().getTableById(tableId);
-        if (table != null && split.length == 2) {
-          String versionId = split[1];
-          if(!StringUtils.isEmpty(versionId)) {
-            version = table.getVersion(versionId);
-          }
-          game = client.getGameService().getGameByVpsTable(table, version);
-        }
+      if (!StringUtils.isEmpty(tournamentTable.getVpsVersionId())) {
+        version = table.getVersion(tournamentTable.getVpsVersionId());
+        game = client.getGameService().getGameByVpsTable(table, version);
       }
 
-      TournamentTreeModel model = new TournamentTreeModel(tournament, game, table, version);
+      TournamentTreeModel model = new TournamentTreeModel(tournament, game, tournamentTable, table, version);
       tournamentNode.getChildren().add(new TreeItem<>(model));
     }
     return tournamentNode;
   }
 
-  public TournamentTreeModel(Tournament tournament, GameRepresentation game, VpsTable vpsTable, VpsTableVersion vpsTableVersion) {
+  public TournamentTreeModel(Tournament tournament, GameRepresentation game, TournamentTable tournamentTable, VpsTable vpsTable, VpsTableVersion vpsTableVersion) {
     this.tournament = tournament;
+    this.tournamentTable = tournamentTable;
     this.vpsTable = vpsTable;
     this.vpsTableVersion = vpsTableVersion;
     this.game = game;
@@ -67,6 +61,14 @@ public class TournamentTreeModel {
       ScoreSummaryRepresentation summary = Studio.client.getGameService().getGameScores(game.getId());
       highscoreAvailable = !StringUtils.isEmpty(summary.getRaw());
     }
+  }
+
+  public TournamentTable getTournamentTable() {
+    return tournamentTable;
+  }
+
+  public boolean isTournamentNode() {
+    return this.tournamentTable == null;
   }
 
   public GameRepresentation getGame() {